requirements.txt
asgiref==3.2.10
cffi==1.14.2
cryptography==3.1
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
gunicorn==20.0.4
Markdown==3.2.2
psycopg2==2.8.5
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.2.0
